spryker/demoshop#333 removed unneeded bundles
