--- src/Pyz/Yves/CmsExporter/Communication/Plugin/PageResourceCreator.php
@@ -1,26 +1,0 @@
-<?php
-
-/**
- * (c) Spryker Systems GmbH copyright protected
- */
-
-namespace Pyz\Yves\CmsExporter\Communication\Plugin;
-
-use SprykerEngine\Yves\Kernel\Communication\AbstractPlugin;
-use Pyz\Yves\CmsExporter\Communication\CmsExporterDependencyContainer;
-
-/**
- * @method CmsExporterDependencyContainer getDependencyContainer()
- */
-class PageResourceCreator extends AbstractPlugin
-{
-
-    /**
-     * @return PageResourceCreator
-     */
-    public function createPageResourceCreator()
-    {
-        return $this->getDependencyContainer()->createPageResourceCreator();
-    }
-
-}
