+++ src/Pyz/Yves/Product/Communication/Builder/FrontendProductBuilderInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * (c) Spryker Systems GmbH copyright protected
+ */
+
+namespace Pyz\Yves\Product\Communication\Builder;
+
+use Pyz\Yves\Product\Communication\Model\AbstractProduct;
+
+/**
+ * Interface FrontendProductBuilderInterface
+ */
+interface FrontendProductBuilderInterface
+{
+
+    /**
+     * @param array $productData
+     *
+     * @return AbstractProduct
+     */
+    public function buildProduct(array $productData);
+
+}
