+++ src/Pyz/Yves/Redirect/Communication/RedirectDependencyContainer.php
@@ -1,0 +1,28 @@
+<?php
+
+/**
+ * (c) Spryker Systems GmbH copyright protected
+ */
+
+namespace Pyz\Yves\Redirect\Communication;
+
+use SprykerEngine\Yves\Kernel\Communication\AbstractCommunicationDependencyContainer;
+use Pyz\Yves\Redirect\Communication\ResourceCreator\RedirectResourceCreator;
+
+/**
+ * @method Redirect getFactory()
+ */
+class RedirectDependencyContainer extends AbstractCommunicationDependencyContainer
+{
+
+    /**
+     * @return RedirectResourceCreator
+     */
+    public function createRedirectResourceCreator()
+    {
+        return new RedirectResourceCreator(
+            $this->getLocator()
+        );
+    }
+
+}
