--- src/Pyz/Yves/RedirectExporter/Communication/RedirectExporterDependencyContainer.php
@@ -1,28 +1,0 @@
-<?php
-
-/**
- * (c) Spryker Systems GmbH copyright protected
- */
-
-namespace Pyz\Yves\RedirectExporter\Communication;
-
-use SprykerEngine\Yves\Kernel\Communication\AbstractCommunicationDependencyContainer;
-use Pyz\Yves\RedirectExporter\Communication\ResourceCreator\RedirectResourceCreator;
-
-/**
- * @method RedirectExporter getFactory()
- */
-class RedirectExporterDependencyContainer extends AbstractCommunicationDependencyContainer
-{
-
-    /**
-     * @return RedirectResourceCreator
-     */
-    public function createRedirectResourceCreator()
-    {
-        return new RedirectResourceCreator(
-            $this->getLocator()
-        );
-    }
-
-}
